@@ -1,14 +1,19 @@
-import Vue from 'vue';
-import 'material-design-icons-iconfont/dist/material-design-icons.css';
-import Vuetify from 'vuetify/lib';
+import 'vuetify/styles';
+import { createVuetify } from 'vuetify';
+import * as components from 'vuetify/components';
+import * as directives from 'vuetify/directives';
+import { md } from 'vuetify/iconsets/md';
 
-Vue.use(Vuetify, {
-  iconfont: 'md',
-});
-
-export default new Vuetify({
+export default createVuetify({
+  components,
+  directives,
+  icons: {
+    defaultSet: 'md',
+    sets: {
+      md,
+    },
+  },
   theme: {
-    dark: false,
-    iconfont: 'md',
-  }
+    defaultTheme: 'light',
+  },
 });
