@@ -1,21 +1,21 @@
-// import Vue from 'vue'
-import Vuetify from 'vuetify'
-import { mount, createLocalVue } from '@vue/test-utils'
+import { mount } from '@vue/test-utils'
 import { expect } from 'chai'
+import { createVuetify } from 'vuetify'
 import TreeItem from '@/components/TreeItem.vue'
-
-const localVue = createLocalVue()
 
 describe('TreeItem.vue', () => {
   let vuetify
+
   beforeEach(() => {
-    vuetify = new Vuetify()
+    vuetify = createVuetify()
   })
+
   it('renders', () => {
     const wrapper = mount(TreeItem, {
-      localVue,
-      vuetify,
-      propsData: {
+      global: {
+        plugins: [vuetify],
+      },
+      props: {
         item: { isRoot: true },
       },
     })
